Merge develop into main (#7)
* Notifications for news

* Send sign up confirmation replaced with resend

* Change name from variable to access to public app url

* Resend API error handler and Key validation

The API key should be validated at initialization. If RESEND_API_KEY is undefined or empty, the Resend instance will be created with an invalid configuration, but the error won't be caught until the first email send attempt.
The sendEmail function should handle errors from the Resend API. Currently, if resend.emails.send throws an error, it will propagate uncaught to the caller.